Create DLLD_29_09_22_02.cpp
Create a node with a name as Node. A Node object has an integer data field, and a Node forwd pointer, pointing to forward node (i.e.: the next node in a list) and a Node back pointer ,pointing to previous node(i.e,: the previous node in a list). In this node builder is inserting N house numbers at the end of list. A node is used to store N House Numbers .Create a program to help a builder to remove the last house number from colony and then print the list in both forward and backward direction.If N>=5 && N<15 condition is not satisfied then print Invalid Input.