Create Stack using Two queue.cpp
Construct a stack using two queues (q1, q2), you need to simulate the stack operations by using queue operations. You are required to complete the three methods push() which takes an integer 'x' as input denoting the element to be pushed into the stack, pop() which poped out from the stack and display() which display the resultant stack(-1 if the stack is empty).