Create QueueImplementation.cpp
Write a program to implement the queue of characters in the array of size 7 in circular order. Take 2 variables ‘front’ and ‘rear’ contains the indexed of the first and last elements of the queue. Where the ‘-1’ in front and rear will represent the empty queue. Your program should be user interactive where 1 will be for insertion and 2 will be for deletion of elements. When the user will enter any other option, it should display the value of the front and rear.