Create 12 QueueImplementation.cpp
Write a program to create the queue of N elements in increasing order. Size of the queue should be greater than 2 and less than equal to 10. All the elements of the queue should be in increasing order. Element will not inserted in the queue if it will not follow the above condition.