Create 15 Queue even element.cpp
Write a program to create the queue of N elements . Size of the queue should be greater than 2 and less than equal to 15. If the size of the queue will not be in the mentioned range then it should display the message “Invalid Queue range" with taking any elements for input.And if there is no even element present then display "No even element is there".