Create 19 Queue_ Head's total fearfulness.cpp
All the employee (except the Head ) are standing in queue to submit their bills. The employee have different seniority. In all there are N employee of K different seniority levels. These are given to you in an array, where A1, A2, ..., AN denote the seniority of employees in the queue. AN denotes front of the queue and A1 denotes end of the queue. Head gets an interesting thought past his head. He begins to think what if every employee starting from the end of the queue begins to delegate his job of submitting bills to a employee least ahead of him in the queue but junior to him. The Head’s fearfulness of this scenario is f = i2 - i1 + 1, where i1 is the index of employee in queue and i2 is the index of the junior employee. Head's total fearfulness of chaos is the product of all the fearfulness in Head’s mind. Note if a employee has no junior ahead of him/her in the queue then Head’s fearfulness for this employee is 1. You are required to find the Head's total fearfulness given an arrangement of employees in a queue. Since this number can be quite large output it modulo 1000000007.