Create 25 Recursion_ Find Middle of List.cpp
given a singly linked list, find the middle of the linked list. For example, if the given linked list is A->B->C->D->E then the output should be C. If there are even nodes, then there would be two middle nodes, we need to print the first middle element. For example, if the given linked list is A->B->C->D->E->F then the output should be C.
Recursively...
void middle(llist *slow, llist *fast)
{
    if(slow==NULL)
    {
        return;
    }
    if(fast->next==NULL)
    {
        cout<<slow->data;
        return;
    }
    if(fast->next->next==NULL)
    {
        cout<<slow->data;
        return;
    }
    if(slow!=NULL || fast->next!=NULL || fast->next->next!=NULL)
    {
        middle(slow->next,fast->next->next);
    }
    
}