Create 26 Insertion in Circular Linked List Recursively.cpp
Ajay got a new problem in his subject. He has given a Circular Singly Linklist in which every node contains a pointer ‘next’ which points to the next node in the list. Last node ‘next’ pointer points to the first node in the linked list. Every node has a ‘data’ part that contains some integer positive. You have to insert an integer Z in the linked list at the given position N.